requirements.txt
torch
torchvision
gradio
Pillow
scikit-learn
tqdm
matplotlib